Create KL7
KL7 is one of the best hotels in patna. KL7 Hotel & Banquets, the most customer-friendly hotel, provides you with all the basic and special facilities to give you a comfortable stay. This hotel maintains the highest standard of international hospitality in exquisite surroundings to give a relaxed and tension-free environment to the guests.